Refactor GameStartScreen to a proper '.ui.' place
